Make it possible to create internships and apply positions

+++ plugins/IctCollege/CoordinatorApprovedSelector/src/Controller/InternshipApplicationsController.php
@@ -3,6 +3,7 @@
 namespace IctCollege\CoordinatorApprovedSelector\Controller;
 
 use Cake\Event\Event;
+use Cake\Network\Exception\BadRequestException;
 
 /**
  * @property \IctCollege\CoordinatorApprovedSelector\Model\Table\InternshipApplicationsTable InternshipApplications
@@ -14,10 +15,12 @@ class InternshipApplicationsController extends AppController
     {
         $application = $this->InternshipApplications->find()->where([
             'position_id' => $this->request->data('position_id'),
-            'student_id' => $this->Auth->user('student_id')
+            'student_id' => $this->Auth->user('student_id'),
         ])->firstOrFail();
 
-        ;
+        if ($application->accepted_coordinator) {
+            throw new BadRequestException(__('Accepted application can not be removed'));
+        }
 
         $this->set('success', (bool)$this->InternshipApplications->delete($application));
         $this->set('_serialize', ['success']);
@@ -52,6 +55,12 @@ class InternshipApplicationsController extends AppController
         /* @var \Cake\ORM\Entity $entity */
         $entity = $event->subject()->entity;
 
-        $entity->student_id = $this->Auth->user('student_id');
+        $studentId = $this->Auth->user('student_id');
+        $internship = $this->InternshipApplications->Periods->Internships->find('active', [
+            'student' => $studentId
+        ])->firstOrFail();
+
+        $entity->period_id = $internship->period_id;
+        $entity->student_id = $studentId;
     }
 }
